feat(session): add SessionTimer component for real-time session countdown and refresh
