+++ src/components/DashboardLayout.tsx
@@ -7,6 +7,7 @@ import { useAuth } from "@/lib/auth";
 import Sidebar from "@/components/Sidebar";
 import ClockInButton from "@/components/ClockInButton";
 import Modal from "@/components/Modal";
+import SessionTimer from "@/components/SessionTimer";
 import { SESSION_CONFIG } from "@/config";
 
 interface DashboardLayoutProps {
@@ -29,7 +30,6 @@ export default function DashboardLayout({
   const [isOpen, setIsOpen] = useState(false);
   const [showSessionModal, setShowSessionModal] = useState(false);
   const [remainingSessionTime, setRemainingSessionTime] = useState(0);
-  const [isRefreshing, setIsRefreshing] = useState(false);
   const [showIdleWarning, setShowIdleWarning] = useState(false);
   const { data: session, update } = useSession();
 
@@ -38,39 +38,6 @@ export default function DashboardLayout({
       router.push("/login");
     }
   }, [user, isLoading, router]);
-
-  // Manual session refresh
-  const refreshSession = useCallback(async () => {
-    setIsRefreshing(true);
-    await update();
-    setTimeout(() => setIsRefreshing(false), 500);
-  }, [update]);
-
-  // Auto refresh session on user activity
-  useEffect(() => {
-    const handleUserActivity = () => {
-      // Only refresh if we're below auto refresh threshold
-      if (
-        remainingSessionTime <= SESSION_CONFIG.autoRefreshThreshold &&
-        remainingSessionTime > 0
-      ) {
-        refreshSession();
-      }
-    };
-
-    // Listen for user activity events
-    window.addEventListener("mousemove", handleUserActivity);
-    window.addEventListener("mousedown", handleUserActivity);
-    window.addEventListener("keypress", handleUserActivity);
-    window.addEventListener("scroll", handleUserActivity);
-
-    return () => {
-      window.removeEventListener("mousemove", handleUserActivity);
-      window.removeEventListener("mousedown", handleUserActivity);
-      window.removeEventListener("keypress", handleUserActivity);
-      window.removeEventListener("scroll", handleUserActivity);
-    };
-  }, [remainingSessionTime, refreshSession]);
 
   // Real-time session timer for debugging
   useEffect(() => {
@@ -169,44 +136,8 @@ export default function DashboardLayout({
                 <div className="flex items-center gap-4">
                   <ClockInButton />
 
-                  {/* Session Timer (Debug) */}
-                  <div className="flex items-center gap-2">
-                    <div
-                      className={`px-3 py-1 rounded-md font-mono text-sm font-bold transition-all ${
-                        isRefreshing
-                          ? "bg-blue-100 text-blue-700 dark:bg-blue-900 dark:text-blue-300"
-                          : remainingSessionTime > 10
-                            ? "bg-green-100 text-green-700 dark:bg-green-900 dark:text-green-300"
-                            : remainingSessionTime > 5
-                              ? "bg-yellow-100 text-yellow-700 dark:bg-yellow-900 dark:text-yellow-300"
-                              : "bg-red-100 text-red-700 dark:bg-red-900 dark:text-red-300 animate-pulse"
-                      }`}
-                    >
-                      ⏱ {isRefreshing ? "↻" : remainingSessionTime}s
-                    </div>
-
-                    {/* Manual Refresh Button */}
-                    <button
-                      onClick={refreshSession}
-                      disabled={isRefreshing}
-                      className="p-1 rounded hover:bg-zinc-100 dark:hover:bg-zinc-700 transition-colors disabled:opacity-50"
-                      title="Refresh Session"
-                    >
-                      <svg
-                        className={`w-4 h-4 text-zinc-600 dark:text-zinc-400 ${isRefreshing ? "animate-spin" : ""}`}
-                        fill="none"
-                        stroke="currentColor"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-                        />
-                      </svg>
-                    </button>
-                  </div>
+                  {/* Session Timer */}
+                  <SessionTimer />
 
                   <div className="h-8 w-px bg-zinc-200 dark:bg-zinc-700" />
                   <div className="relative">
@@ -335,7 +266,7 @@ export default function DashboardLayout({
                         </p>
                         <button
                           onClick={async () => {
-                            await refreshSession();
+                            await update();
                             setShowIdleWarning(false);
                           }}
                           className="px-6 py-2 bg-blue-600 hover:bg-blue-700 text-white rounded-md font-medium transition-colors"
